move line interactable to its own file
