+++ lib/src/deriv_chart/interactive_layer/interactable_drawings/interactable_drawing.dart
@@ -1,0 +1,122 @@
+import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tool_config.dart';
+import 'package:flutter/widgets.dart';
+
+import '../../chart/data_visualization/chart_data.dart';
+import '../../chart/data_visualization/models/animation_info.dart';
+import '../interactable_drawing_custom_painter.dart';
+
+/// Represents the current state of a drawing tool on the chart.
+///
+/// The state determines how the drawing tool is rendered and how it responds
+/// to user interactions. Different states trigger different visual appearances
+/// and interaction behaviors.
+enum DrawingToolState {
+  /// Default state when the drawing tool is displayed on the chart
+  /// but not being interacted with.
+  normal,
+
+  /// The drawing tool is currently selected by the user. Selected tools
+  /// typically show additional visual cues like handles or a glowy effect
+  /// to indicate they can be manipulated.
+  selected,
+
+  /// The user's pointer is hovering over the drawing tool but hasn't
+  /// selected it yet. This state can be used to provide visual feedback
+  /// before selection.
+  hovered,
+
+  /// The drawing tool is in the process of being created/added to the chart.
+  /// In this state, the tool captures user inputs (like taps) to define
+  /// its shape and position.
+  adding,
+
+  /// The drawing tool is being actively moved or resized by the user.
+  /// This state is active during drag operations when the user is
+  /// modifying the tool's position.
+  dragging,
+}
+
+/// The class that will be generated by the drawing tool config instance when
+/// they are created or the saved ones that are loaded from storage.
+/// The information from this class (its subclasses) will be used to draw the
+/// tool on the chart.
+/// It will keep the latest state of the drawing tool as the user interacts
+/// with the tools in the runtime.
+/// During the time that user interacts with a tool. by some debounce mechanism
+/// This class will update the config which is supposed to be saved in the storage.
+abstract class InteractableDrawing<T extends DrawingToolConfig> {
+  /// Initializes [InteractableDrawing].
+  InteractableDrawing({required this.config});
+
+  static const double _hitTestMargin = 32;
+
+  /// The margin for hit testing.
+  double get hitTestMargin => _hitTestMargin;
+
+  /// The drawing tool config.
+  final T config;
+
+  /// Returns the updated config.
+  T getUpdatedConfig();
+
+  /// Returns `true` if the drawing tool is hit by the given offset.
+  bool hitTest(Offset offset, EpochToX epochToX, QuoteToY quoteToY);
+
+  /// The tap event that is called when the [InteractableDrawing] is in adding
+  /// state.
+  ///
+  /// the drawing can use the tap to capture and create the coordinates required
+  /// for its shape.
+  ///
+  /// [onDone] is a callback that should be called when the drawing is done.
+  void onCreateTap(
+    TapUpDetails details,
+    EpochFromX epochFromX,
+    QuoteFromY quoteFromY,
+    EpochToX epochToX,
+    QuoteToY quoteToY,
+    VoidCallback onDone,
+  ) {}
+
+  /// Called when the drawing tool dragging is started.
+  void onDragStart(
+    DragStartDetails details,
+    EpochFromX epochFromX,
+    QuoteFromY quoteFromY,
+    EpochToX epochToX,
+    QuoteToY quoteToY,
+  ) {}
+
+  /// Called when the drawing tool is dragged and updates the drawing position
+  /// properties based on the dragging [details].
+  ///
+  /// Each drawing will know how to handle and update itself accordingly based
+  /// on where the dragging position is like if it's dragging a point or a line
+  /// of the tool.
+  void onDragUpdate(
+    DragUpdateDetails details,
+    EpochFromX epochFromX,
+    QuoteFromY quoteFromY,
+    EpochToX epochToX,
+    QuoteToY quoteToY,
+  );
+
+  /// Called when the drawing tool dragging is ended.
+  void onDragEnd(
+    DragEndDetails details,
+    EpochFromX epochFromX,
+    QuoteFromY quoteFromY,
+    EpochToX epochToX,
+    QuoteToY quoteToY,
+  ) {}
+
+  /// Paints the drawing tool on the chart.
+  void paint(
+    Canvas canvas,
+    Size size,
+    EpochToX epochToX,
+    QuoteToY quoteToY,
+    AnimationInfo animationInfo,
+    GetDrawingState getDrawingState,
+  );
+}
