+++ lib/src/deriv_chart/interactive_layer/interactive_layer_base.dart
@@ -1,7 +1,7 @@
 import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tool_config.dart';
 
 import '../chart/data_visualization/chart_data.dart';
-import 'interactable_drawing.dart';
+import 'interactable_drawings/interactable_drawing.dart';
 import 'interactive_states/interactive_state.dart';
 
 /// The interactive layer base class interface.
