+++ lib/src/deriv_chart/interactive_layer/interactive_states/interactive_normal_state.dart
@@ -2,7 +2,7 @@ import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tool_config.dar
 import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 
-import '../interactable_drawing.dart';
+import '../interactable_drawings/interactable_drawing.dart';
 import 'Interactive_hover_state.dart';
 import 'interactive_selected_tool_state.dart';
 import 'interactive_state.dart';
